requirements.txt
aiogram==2.25.1
python-pptx
openai
python-dotenv
pydantic
requests
